requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
arxiv==2.1.3
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.7
decorator==5.1.1
distro==1.9.0
dnspython==2.7.0
e2b==1.0.2
e2b-code-interpreter==1.0.1
email_validator==2.2.0
executing==2.1.0
faiss-cpu==1.9.0
fastapi==0.115.3
fastapi-cli==0.0.5
fastembed==0.4.1
fastjsonschema==2.20.0
feedparser==6.0.11
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
greenlet==3.1.1
groq==0.11.0
grpcio==1.67.1
grpcio-status==1.67.0
grpcio-tools==1.67.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
ipython-genutils==0.2.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
langchain==0.3.6
langchain-community==0.3.4
langchain-core==0.3.14
langchain-experimental==0.3.2
langchain-google-genai==2.0.1
langchain-groq==0.2.0
langchain-text-splitters==0.3.0
langgraph==0.2.41
langgraph-checkpoint==2.0.2
langgraph-sdk==0.1.35
langsmith==0.1.138
loguru==0.7.2
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==4.1.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.26.4
onnx==1.17.0
onnxruntime==1.19.2
orjson==3.10.10
packaging==24.1
pandas==2.2.3
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pillow==10.4.0
pinecone==5.3.1
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
plotly==5.24.1
plotly-express==0.4.1
portalocker==2.10.1
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
protoc-gen-openapiv2==0.0.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py_rust_stemmers==0.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.6.0
pydantic_core==2.23.4
Pygments==2.18.0
PyMuPDF==1.24.12
pyparsing==3.2.0
PyStemmer==2.2.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.16
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qdrant-client==1.12.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.3
rpds-py==0.20.0
rsa==4.9
scipy==1.14.1
seaborn==0.13.2
setuptools==75.3.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.0
statsmodels==0.14.4
sympy==1.13.3
tavily-python==0.5.0
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.20.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
yarl==1.16.0
